feat(Participant): ACE 계산 로직 추가
