+++ README.md
@@ -40,7 +40,7 @@
 
 - [ ] 모든 플레이어는 카드를 받을 때마다 점수를 추가해서 저장한다.
 - [ ] J, Q, K는 10으로 계산한다.
-- [ ] Ace는 11로 계산하되, 합계가 21이 넘는 경우 1로 계산한다.
+- [x] Ace는 11로 계산하되, 합계가 21이 넘는 경우 1로 계산한다.
 - [ ] 모든 딜러와 플레이어는 점수가 21을 초과하면 `Bust` 된다.
 
 #### 결과 출력
