+++ src/main/java/domain/Player.java
@@ -2,24 +2,9 @@ package domain;
 
 import java.util.List;
 
-public class Player {
-    private final Name name;
-    private final List<Card> cards;
+public class Player extends Participant {
 
     public Player(final Name name, final List<Card> cards) {
-        this.name = name;
-        this.cards = cards;
-    }
-
-    public void receiveCard(Card card) {
-        this.cards.add(card);
-    }
-
-    public int calculateScore() {
-        int score = 0;
-        for (Card card : cards) {
-            score += card.getScore();
-        }
-        return score;
+        super(name, cards);
     }
 }
